Search by Mood range, UI component added
Use slide bar to select mood range and filter according to the
range.Compatible with keyword search at the same time.

Mood representation visualized. Heart is blue if rating is below 3.
Heart is red if rating is greater or equal to 3.

Divs hid and show controlled with button to avoid to be too crowd on
the screen.